refactor(app): #373 Phase 1-4 use-team-management.ts に teams / TeamHub を切り出し
App.tsx 2107 -> 1656 行 (-451)。teams / team-history / TeamHub 接続情報・
doCloseTeam・handleResumeTeam・get*Args 系を新規 hook
src/renderer/src/lib/hooks/use-team-management.ts に集約。

新規ファイル:
- src/renderer/src/lib/team-prompts.ts (88 行)
  ROLE_DESC / ROLE_ORDER / generateTeamSystemPrompt / generateTeamAction
  を純粋関数 module として独立 (テスト容易性 + Canvas 側との将来共有)。
- src/renderer/src/lib/hooks/use-team-management.ts (524 行)
  state / handler / IPC を集中管理。

移管:
- state: teams / teamHistoryEntries / teamHubInfo / spawnStaggerTimers /
  teamHistory debounce 一式 (pending Map / flushTimer)
- effect: アンマウント flush + spawn timer cleanup / TeamHub 接続情報の
  起動時取得 / projectRoot 変更時の team-history 自動ロード
- handler: doCloseTeam / handleCloseLeaderOnly / handleResumeTeam /
  handleDeleteTeamHistory / handleTerminalSessionId /
  persistTerminalCustomLabel / saveTeamHistory / refreshTeamHistory /
  flushTeamHistoryNow / clearSpawnTimers
- launch helpers: getTerminalArgs / getCodexInstructions /
  getTerminalEnv / getRolePrompt (TerminalView の props として
  hook 戻り値からそのまま渡せる)

Phase 1-1 〜 1-3 と同じ流儀:
- opts (projectRoot / showToast / Phase 1-3 hook 戻り値ブリッジ) を
  optsRef.current = opts で毎 render 更新し、内部 useCallback の
  deps を最小化
- useT / useSettingsValue は hook 内で直接呼ぶ
- resetForProjectSwitch を expose し projectSwitchedRef から呼ぶ流儀

useTerminalTabs の API 整理:
- opts.teams を撤廃 (旧用途の standaloneTabList / teamGroupList が
  実は JSX 側で未使用 = dead code だったため削除)
- result から standaloneTabList / teamGroupList を削除
- これにより useTerminalTabs ↔ useTeamManagement の TDZ 循環を解消
  (teams を opts で渡す必要がなくなった)

App.tsx での hook 接続:
- closeTeamRef ブリッジは 1 本だけ残置 (useTerminalTabs.opts.closeTeam
  → useTeamManagement.doCloseTeam の唯一の逆方向参照)
- doCloseTeam の本体は hook 内に移動。App.tsx 側は forwarder のみ
  (Phase 1-3 から大幅に縮小)
- projectSwitchedRef.current で setTeams([]) を呼んでいた箇所を
  resetTeamsForProjectSwitch() に置換

検証:
- npm run typecheck: green
- cargo check --manifest-path src-tauri/Cargo.toml: green
- cargo clippy --no-deps: 既存ベースライン 15 件のまま (新規警告ゼロ)

累計削減 (Phase 1-1 〜 1-4): -1171 行 (#373 全体目標 -1995, 進捗 59%)。

Refs #373
