+++ src/renderer/src/lib/team-prompts.ts
@@ -1,0 +1,88 @@
+import type { Team, TeamRole } from '../../../types/shared';
+import {
+  getRoleDisplayLabel,
+  type TerminalTab
+} from './hooks/use-terminal-tabs';
+
+/** ロール別の短い説明（チームプロンプト内で使用、leader 以外は動的ロール由来）。 */
+export const ROLE_DESC: Record<TeamRole, string> = {
+  leader: '全体の調整・指示・タスク割り振り'
+};
+
+/**
+ * ロスター表示用の固定順。leader を最優先に、それ以外は登場順。
+ * vibe-team のロールは Leader が動的に作成するため、固定リスト化はしない。
+ */
+export const ROLE_ORDER: Record<string, number> = {
+  leader: 0
+};
+
+/** チームのシステムプロンプト（--append-system-prompt 用） */
+export function generateTeamSystemPrompt(
+  tab: TerminalTab,
+  allTabs: TerminalTab[],
+  team: Team | null
+): string | undefined {
+  if (!tab.role || !tab.teamId || !team) return undefined;
+
+  const teamTabs = allTabs
+    .filter((t) => t.teamId === tab.teamId)
+    .slice()
+    .sort((a, b) => {
+      const ra = ROLE_ORDER[a.role ?? ''] ?? 99;
+      const rb = ROLE_ORDER[b.role ?? ''] ?? 99;
+      if (ra !== rb) return ra - rb;
+      return a.agentId.localeCompare(b.agentId);
+    });
+  const roster = teamTabs
+    .map((t) => {
+      const agent = t.agent === 'claude' ? 'Claude Code' : 'Codex';
+      const you = t.id === tab.id ? ' ← あなた' : '';
+      const roleLabel = getRoleDisplayLabel(t, allTabs);
+      return `${roleLabel || 'member'}(${agent})${you}`;
+    })
+    .join(', ');
+
+  const mcpTools =
+    'MCP vibe-team ツール: team_recruit(role_id,engine,label?,description?,instructions?) / team_dismiss / team_send(to,message) / team_read / team_info / team_status / team_assign_task(assignee,description) / team_get_tasks / team_update_task / team_list_role_profiles。' +
+    'team_send/team_assign_task は相手のプロンプトにリアルタイム注入される。受信時は [Team ← <role>] プレフィックス付きで届く。';
+
+  if (tab.role === 'leader') {
+    return (
+      `あなたはチーム「${team.name}」のLeader。構成: ${roster}。${mcpTools}\n` +
+      `【絶対遵守ルール — 外部ファイルを読む前に先に従うこと】\n` +
+      `1. ユーザーから最初の指示が来るまで何もせず待機する。自分からプロジェクト調査やファイル読みを開始しない。\n` +
+      `2. ユーザー指示が届いたら、計画して委譲する。Read / Edit / Write / Bash / Grep / Glob などの作業系ツールを Leader 自身が呼んで実作業をしてはいけない。Leader の仕事は計画・委譲・レビュー。\n` +
+      `【チーム編成とタスク委譲の使い分け】\n` +
+      `(a) vibe-team (基本・可視化): team_recruit + team_assign_task を使うとキャンバス上にメンバーが視覚的に配置される。「チームを作って」「採用して」と言われたときや、通常のタスク委譲はこれを既定で使う。\n` +
+      `(b) Claude Code Native Agent Teams (Task / dispatch_agent / general-purpose / Explore): ユーザーから「裏で Agent Teams を使って」「サブエージェントに任せて」と明示指示されたとき、またはキャンバスに表示するまでもない大量ファイル検索 / 裏側の単純並列タスクを Leader 自身の判断で行うときのみ使用。通常の委譲を勝手にこっちに振り替えない。\n` +
+      `3. team_recruit は「ロール設計＋採用」を 1 コールで行う。新規ロール作成時の必須引数: role_id (snake_case), label, description, instructions, engine。` +
+      `既存ロール (hr や自分が作成済みの role_id) の再採用は role_id + engine だけで OK。\n` +
+      `4. 3 名以上必要なときは、まず team_recruit({role_id:"hr", engine:"claude"}) で HR を採用し、team_send("hr", "採用してほしい: ...") で一括採用を委譲する。\n` +
+      `5. チームが揃ったら team_assign_task で割り振り、結果は [Team ← <role>] で届くので都度レビュー、追指示は team_send で行う。\n` +
+      `6. 【長文ペイロード・ルール】team_recruit.instructions / team_send.message / team_assign_task.description は bracketed paste で配送されるので改行入り YAML / code / リストも ~32 KiB まではそのままインラインで OK。32 KiB を超える本文のみ Write で .vibe-team/tmp/<short_id>.md に書き出してから引数には「サマリ + パス」を渡す (Hub が 32 KiB 超を拒否)。\n` +
+      `設計思想や応用パターンの詳細は .claude/skills/vibe-team/SKILL.md を Read ツールで参照可 (補助情報、必須ではない)。`
+    );
+  }
+
+  // leader 以外: 役割の詳細はロールプロファイル (動的生成可能) 側で管理されるため、
+  // ここでは固定の汎用文だけを返す。IDE 旧仕様の fallback。Canvas 側は AgentNodeCard が
+  // renderSystemPrompt() で動的ロール instructions を含むプロンプトを組み立てる。
+  const roleDesc = ROLE_DESC[tab.role] ?? `${tab.role}としての担当作業`;
+  return (
+    `あなたはチーム「${team.name}」の${tab.role}。役割:${roleDesc}。構成: ${roster}。${mcpTools}\n` +
+    `【絶対ルール】\n` +
+    `1. 指示が [Team ← leader] (または [Team ← <role>]) で届くまで何もしない。自発的な調査・コード変更は禁止。\n` +
+    `2. 指示が届いたら作業を完遂し、直後に team_send('leader', "完了報告: ...") で簡潔に結果を返す。\n` +
+    `3. 報告後は静かなアイドル状態に戻る。ポーリング・「承認待ち」表示・自発的な追加質問は禁止。次の指示は [Team ← ...] で自動的に届く。\n` +
+    `4. 自分から他メンバーにタスクを割り振ってはいけない (それは Leader の仕事)。\n` +
+    `5. 【長文ペイロード・ルール】team_send は bracketed paste で配送されるので改行入りの内容も ~32 KiB まではそのまま OK。それを超える場合のみ Write で .vibe-team/tmp/<short_id>.md に書き出してパスを渡す。`
+  );
+}
+
+/** 短いアクション指示（initialMessage 用）。
+ *  チーム所属タブは全員「待機」が基本方針なので何も送らない。
+ *  Leader はユーザーからの最初の指示を待ち、メンバーは Leader からの注入を待つ。 */
+export function generateTeamAction(_tab: TerminalTab): string | undefined {
+  return undefined;
+}
